+++ client/src/components/layouts/MismatchOverlay/index.tsx
@@ -1,0 +1,10 @@
+export { default as MismatchOverlay } from './MismatchOverlay';
+export { default as MismatchContent } from './MismatchContent';
+export { default as MismatchSuggestion } from './MismatchSuggestion';
+
+// Re-export types
+export type {
+  MismatchOverlayProps,
+  MismatchContentProps,
+  MismatchSuggestionProps,
+} from '../../../types/mismatchTypes';